django/clase/noticias_upsa/requirements.txt
django==3.1.7
djangorestframework==3.12.2